+++ app/app.js
@@ -1,5 +1,5 @@
-import { start } from './scripts/start';
 import './branding/js/branding.js';
+import { start } from './scripts/start';
 
 window.$ = window.jQuery = require('./vendor/jquery-1.12.0');
 start();
